refactor: move session utils to `@logto/node`
