+++ packages/client/src/shim.ts
@@ -18,7 +18,7 @@ export {
   getOrganizationIdFromUrn,
 } from '@logto/js';
 export * from './errors.js';
-export type { Storage, StorageKey, ClientAdapter } from './adapter/index.js';
+export type { Storage, StorageKey, ClientAdapter, JwtVerifier } from './adapter/index.js';
 export { PersistKey, CacheKey } from './adapter/index.js';
 export { createRequester } from './utils/index.js';
 export * from './types/index.js';
