+++ packages/node/tsconfig.build.json
@@ -3,5 +3,6 @@
   "include": [
     "src",
     "edge"
-  ]
+  ],
+  "exclude": ["src/**/*.test.ts", "edge/**/*.test.ts"]
 }
